replace schedule service with a decorator
