+++ src/services/FileCleaner.ts
@@ -1,25 +1,21 @@
-import { Constant, Inject, Service } from "@tsed/di";
+import { constant as getFromEnv, Inject, Service } from "@tsed/di";
 import { OnReady } from "@tsed/common";
 import { FileRepo } from "../db/repo/FileRepo.js";
-import { ScheduleService } from "./ScheduleService.js";
 import { filesDir, FileUtils } from "../utils/Utils.js";
 import GlobalEnv from "../model/constants/GlobalEnv.js";
 import fs from "node:fs/promises";
 import { FileUploadModel } from "../model/db/FileUpload.model.js";
 import { FileService } from "./FileService.js";
+import { RunEvery } from "../model/di/decorators/RunEvery.js";
 
 @Service()
 export class FileCleaner implements OnReady {
     public constructor(
         @Inject() private repo: FileRepo,
-        @Inject() private scheduleService: ScheduleService,
         @Inject() private fileUploadService: FileService,
     ) {}
 
-    // default to every hour at :00
-    @Constant(GlobalEnv.FILE_CLEANER_CRON, "0 * * * *")
-    private readonly cronToRun: string;
-
+    @RunEvery(() => getFromEnv(GlobalEnv.FILE_CLEANER_CRON, "0 * * * *"), true)
     public async processFiles(): Promise<void> {
         const allFiles = await this.repo.getExpiredFiles();
         if (allFiles.length === 0) {
@@ -28,17 +24,9 @@ export class FileCleaner implements OnReady {
         await this.fileUploadService.processDelete(allFiles.map(entry => entry.token));
     }
 
-    public $onReady(): void {
-        this.scheduleService.scheduleCronJob(
-            this.cronToRun,
-            async () => {
-                await this.processFiles();
-                await this.sync();
-            },
-            "removeExpiredFiles",
-            this,
-            true,
-        );
+    public async $onReady(): Promise<void> {
+        await this.processFiles();
+        await this.sync();
     }
 
     private async sync(): Promise<void> {
